moved printing functions to tasks
